% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalipseoModelManager.R
\name{CalipseoModelManager}
\alias{CalipseoModelManager}
\title{Calipseo Model Manager}
\format{
R6 class
}
\description{
This R6 class handles connection and SQL metadata management for the Calipseo data model.
It allows listing tables available in the database and loading individual tables as R6 objects.
}
\section{Methods}{

\describe{
  \item{\code{new(pool)}}{Create a new manager using a database connection pool.}
  \item{\code{list_tables()}}{List all tables in the database.}
  \item{\code{load_table(table_name)}}{Load a specific SQL table as a \code{CalipseoTable} object.}
}
}

\examples{
\dontrun{
pool <- pool::dbPool(DBI::dbConnect(...))
manager <- CalipseoModelManager$new(pool)
manager$list_tables()
table_obj <- manager$load_table("my_table")
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalipseoModelManager-new}{\code{CalipseoModelManager$new()}}
\item \href{#method-CalipseoModelManager-list_tables}{\code{CalipseoModelManager$list_tables()}}
\item \href{#method-CalipseoModelManager-load_table}{\code{CalipseoModelManager$load_table()}}
\item \href{#method-CalipseoModelManager-clone}{\code{CalipseoModelManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoModelManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoModelManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoModelManager$new(pool, db_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoModelManager-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoModelManager-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoModelManager$list_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoModelManager-load_table"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoModelManager-load_table}{}}}
\subsection{Method \code{load_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoModelManager$load_table(table_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoModelManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoModelManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoModelManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
