% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalipseoTable.R
\name{CalipseoTable}
\alias{CalipseoTable}
\title{Calipseo SQL Table Handler}
\format{
R6 class
}
\description{
This R6 class represents a SQL table from the Calipseo model.
It provides functions to manage data entries and generate SQL insert statements.
}
\section{Methods}{

\describe{
  \item{\code{new(pool, table_name)}}{Create a table object connected to SQL.}
  \item{\code{TABLE_INFO()}}{Return SQL metadata of the table.}
  \item{\code{FOREIGN_KEYS()}}{Return foreign keys related to this table.}
  \item{\code{ADD_ENTRY(...)}}{Add a new row entry.}
  \item{\code{VIEW_ENTRIES()}}{View current staged entries (data.table).}
  \item{\code{INSERT_STATEMENT()}}{Generate SQL insert statements from staged entries.}
  \item{\code{CLEAR_STATEMENT()}}{Generate SQL to clear the table (TRUNCATE).}
}
}

\examples{
\dontrun{
pool <- pool::dbPool(DBI::dbConnect(...))
table_obj <- CalipseoTable$new(pool, "my_table")
table_obj$ADD_ENTRY(name = "Test", quantity = 42)
table_obj$INSERT_STATEMENT()
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalipseoTable-new}{\code{CalipseoTable$new()}}
\item \href{#method-CalipseoTable-ADD_ENTRY}{\code{CalipseoTable$ADD_ENTRY()}}
\item \href{#method-CalipseoTable-REMOVE_ENTRY}{\code{CalipseoTable$REMOVE_ENTRY()}}
\item \href{#method-CalipseoTable-CLEAR_ENTRIES}{\code{CalipseoTable$CLEAR_ENTRIES()}}
\item \href{#method-CalipseoTable-VIEW_ENTRIES}{\code{CalipseoTable$VIEW_ENTRIES()}}
\item \href{#method-CalipseoTable-TABLE_INFO}{\code{CalipseoTable$TABLE_INFO()}}
\item \href{#method-CalipseoTable-MANDATORY_FIELDS}{\code{CalipseoTable$MANDATORY_FIELDS()}}
\item \href{#method-CalipseoTable-VIEW_DB_TABLE}{\code{CalipseoTable$VIEW_DB_TABLE()}}
\item \href{#method-CalipseoTable-FOREIGN_KEYS}{\code{CalipseoTable$FOREIGN_KEYS()}}
\item \href{#method-CalipseoTable-MAX_INDEX}{\code{CalipseoTable$MAX_INDEX()}}
\item \href{#method-CalipseoTable-INSERT_STATEMENT}{\code{CalipseoTable$INSERT_STATEMENT()}}
\item \href{#method-CalipseoTable-CLEAR_STATEMENT}{\code{CalipseoTable$CLEAR_STATEMENT()}}
\item \href{#method-CalipseoTable-clone}{\code{CalipseoTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$new(con, table_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-ADD_ENTRY"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-ADD_ENTRY}{}}}
\subsection{Method \code{ADD_ENTRY()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$ADD_ENTRY(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-REMOVE_ENTRY"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-REMOVE_ENTRY}{}}}
\subsection{Method \code{REMOVE_ENTRY()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$REMOVE_ENTRY(row_to_remove)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-CLEAR_ENTRIES"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-CLEAR_ENTRIES}{}}}
\subsection{Method \code{CLEAR_ENTRIES()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$CLEAR_ENTRIES()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-VIEW_ENTRIES"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-VIEW_ENTRIES}{}}}
\subsection{Method \code{VIEW_ENTRIES()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$VIEW_ENTRIES()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-TABLE_INFO"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-TABLE_INFO}{}}}
\subsection{Method \code{TABLE_INFO()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$TABLE_INFO()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-MANDATORY_FIELDS"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-MANDATORY_FIELDS}{}}}
\subsection{Method \code{MANDATORY_FIELDS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$MANDATORY_FIELDS()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-VIEW_DB_TABLE"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-VIEW_DB_TABLE}{}}}
\subsection{Method \code{VIEW_DB_TABLE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$VIEW_DB_TABLE()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-FOREIGN_KEYS"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-FOREIGN_KEYS}{}}}
\subsection{Method \code{FOREIGN_KEYS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$FOREIGN_KEYS()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-MAX_INDEX"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-MAX_INDEX}{}}}
\subsection{Method \code{MAX_INDEX()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$MAX_INDEX()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-INSERT_STATEMENT"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-INSERT_STATEMENT}{}}}
\subsection{Method \code{INSERT_STATEMENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$INSERT_STATEMENT(execute = FALSE, file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-CLEAR_STATEMENT"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-CLEAR_STATEMENT}{}}}
\subsection{Method \code{CLEAR_STATEMENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$CLEAR_STATEMENT(execute = FALSE, file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalipseoTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalipseoTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalipseoTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
